fix database

+++ src/main/java/vdtry06/springboot/ecommerce/dto/request/product/ProductRequest.java
@@ -14,7 +14,7 @@ import java.util.Set;
 public class ProductRequest {
     String name;
     String description;
-    Integer availableQuantity;
+    Double availableQuantity;
     Double price;
     Set<String> categoryNames;
 }
